requirements.txt
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
chargily-epay-Django==0.0.2
chargily-epay-python==0.0.2
charset-normalizer==2.0.12
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-storages==1.13.1
djangorestframework==3.14.0
idna==3.4
mod-wsgi==4.9.4
oauthlib==3.2.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
requests==2.27.0
requests-oauthlib==1.3.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
